fix: 마이페이지 댓글 로딩 시 헤더 표시 추가 (#102)
