+++ components/mypage/comments/CommentsBoundary.tsx
@@ -1,11 +1,21 @@
 'use client';
 
+import { useRouter } from 'next/navigation';
+import { ChevronDownIcon } from '@/app/assets/icons';
 import { AsyncBoundary, LoadingSpinner, ErrorFallback } from '@/lib/async-boundary';
 import CommentsSection from './CommentsSection';
 
 export default function CommentsBoundary() {
+  const router = useRouter();
+
   return (
     <div className="flex flex-col gap-5">
+      <div className="flex items-center gap-4 bg-white px-4 py-3 md:px-9 xl:hidden">
+        <button type="button" onClick={() => router.back()} aria-label="뒤로가기">
+          <ChevronDownIcon className="h-5 w-5 rotate-90 text-black" />
+        </button>
+        <h1 className="text-base font-semibold text-black md:text-[18px]">댓글</h1>
+      </div>
       <h1 className="hidden text-lg font-semibold text-black xl:block">댓글</h1>
       <AsyncBoundary
         pendingFallback={<LoadingSpinner className="pt-11.25" />}
